pyproject.toml
[tool.poetry]
name = "kinetik"
version = "0.1.0"
description = ""
authors = ["Aurelien Roy <roy.aurelien@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.3"
pydantic-settings = "^2.7.0"
pyyaml = "^6.0.2"
transitions = "^0.9.2"
msgspec = "^0.18.6"
python-statemachine = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
